주문하기, 주문 완료 화면 개발

+++ packages/domain/lib/entity/login_features.entity.dart
@@ -11,7 +11,7 @@ class LoginFeaturesEntity {
   String? name;
 
   @JsonKey(name: 'point_can_use_limit')
-  int? pointCanUseLimit;
+  double? pointCanUseLimit;
 
   LoginFeaturesEntity({
     this.key,
